implement media and template files in merge-tool
