+++ package.json
@@ -24,7 +24,7 @@
   },
   "dependencies": {
     "@caliorg/a2jdeps": "^7.1.3",
-    "@caliorg/a2jviewer": "^8.0.0",
+    "@caliorg/a2jviewer": "file:../viewer",
     "bit-tabs": "^2.0.0",
     "blueimp-file-upload": "^9.10.1",
     "bootstrap": "^3.4.1",
