+++ src/merge-tool/merge-tool.js
@@ -1,379 +1,12 @@
 import DefineMap from 'can-define/map/map'
 import Component from 'can-component'
-import naturalCompare from 'string-natural-compare/'
 import template from './merge-tool.stache'
 import Guide from '~/src/models/app-state-guide'
-import A2JTemplate from '@caliorg/a2jdeps/models/a2j-template'
-import GuideFiles from '~/src/models/guide-files'
 import { formatPageTextCell, formatBytes } from './helpers/helpers'
+import _findIndex from 'lodash/findIndex'
 
 // guide wrapper that adds functionality for use in the merge tool such as converting to/from the generic recursive accordion data
-export const MergeToolGuide = DefineMap.extend('MergeToolGuide', {
-  gid: {
-    default: undefined
-  },
-  guide: {
-    default: undefined,
-    value ({ lastSet, listenTo, resolve }) {
-      listenTo('loadPromise', (ev, newVal) => {
-        newVal && newVal.then(guide => {
-          const copy = new Guide(guide)
-          // TODO: delete id because this should be a clone?
-          resolve(copy)
-        }) // todo: catch
-      })
-
-      listenTo(lastSet, resolve)
-      resolve(lastSet.get())
-    }
-  },
-  guideFilesPromise: {
-    default: undefined,
-    get () {
-      const gid = this.gid
-      if (gid && gid !== 'a2j') {
-        return GuideFiles.findAll(gid)
-      }
-    }
-  },
-  guideFiles: {
-    default: undefined,
-    get (lastSet, resolve) {
-      if (lastSet) {
-        return lastSet
-      }
-
-      this.guideFilesPromise && this.guideFilesPromise.then(resolve)
-    }
-  },
-
-  // templatesPromise: {
-  //   default: undefined,
-  //   get () {
-  //     const guideId = this.gid
-  //     if (guideId && guideId !== 'a2j') {
-  //       return A2JTemplate.findAll({ guideId })
-  //     }
-  //   }
-  // },
-
-  // templates: {
-  //   default: undefined,
-  //   get (lastSet, resolve) {
-  //     if (lastSet) {
-  //       return lastSet
-  //     }
-
-  //     this.templatesPromise && this.templatesPromise.then(resolve)
-  //   }
-  // },
-  loadPromise: {
-    type: 'any'
-  },
-  load (gid) {
-    // openSelectedGuide(gid) // legacy/A2J_Guides.js
-    // -> calls: ws({ cmd: 'guide', gid: gid }, guideLoaded) // legacy/A2J_AuthorApp.js
-    // guideLoaded is a callback that recieves xml version of the guide from the web service
-    // then it parses it into json via parseXML_Auto_to_CAJA and sets a global 'gGuide'
-    // TODO: similar path to what the main app does, as described in the comments above. ^
-    // then return a promise that resolves to the guide pojo like the following (temporary mock) line
-    // const loadPromise = fetch(new Request('./resources/' + gid + '.json')).then(response => response.json())
-
-    let loadPromise
-
-    if (gid === 'a2j') {
-      const a2jGuide = window.blankGuide()
-      loadPromise = Promise.resolve(a2jGuide)
-    } else {
-      this.gid = gid
-
-      loadPromise = window.fetch('CAJA_WS.php', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8' },
-        body: `cmd=guide&gid=${gid}`
-      })
-        .then(response => response.json())
-        .then(data => window.parseXML_Auto_to_CAJA($(jQuery.parseXML(data.guide))))
-        .catch(err => console.error(err))
-    }
-
-    this.loadPromise = loadPromise
-    return loadPromise
-  },
-
-  /**
-   * @property {Function} mergeTool.ViewModel.prototype.guidePartToSortedArray guidePartToSortedArray
-   * @parent mergeTool.ViewModel
-   *
-   * builds a list in natural sort order from a var or pages object on a guide
-   */
-  guidePartToSortedArray (guidePart, sortByProp = 'name') {
-    const sortedList = []
-
-    Object.keys(guidePart).forEach((partKey) => {
-      sortedList.push(guidePart[partKey])
-    })
-
-    return sortedList.sort((a, b) => naturalCompare.caseInsensitive(a[sortByProp], b[sortByProp]))
-  },
-
-  /**
-   * @property {Array} mergeTool.ViewModel.prototype.define.sortedVariableList sortedVariableList
-   * @parent mergeTool.ViewModel
-   *
-   * sorted variable list for display
-   */
-  sortedVariableList: {
-    get () {
-      const guide = this.guide
-      if (guide && guide.vars) {
-        return this.guidePartToSortedArray(guide.vars)
-      }
-    }
-  },
-
-  sortedPages: {
-    get () {
-      const guide = this.guide
-      if (guide && guide.pages) {
-        return this.guidePartToSortedArray(guide.pages)
-      }
-    }
-  },
-
-  // converts a guide into a generic recursive structure that the checkbox accordion can render
-  get accordionFromGuide () {
-    const accordion = [
-      {
-        label: 'Interview Variables',
-        expanded: true,
-        details: [],
-        children: [], // children are each objects in the same shape as this one
-        value: undefined // value can be anything, it's just data to associate with the item
-      },
-      {
-        label: 'Interview Steps and Pages',
-        expanded: true,
-        details: [],
-        children: [],
-        value: undefined
-      },
-      {
-        label: 'PopUp Pages',
-        expanded: true,
-        details: [],
-        children: [],
-        value: undefined
-      },
-      {
-        label: 'Media Files',
-        expanded: true,
-        details: [],
-        children: [],
-        value: undefined
-      },
-      {
-        label: 'Templates',
-        expanded: true,
-        details: [],
-        children: [],
-        value: undefined
-      }
-    ]
-    // first guide parts
-    const guide = this.guide
-    if (guide) {
-      const vars = this.sortedVariableList
-      accordion[0].children = vars.map(v => {
-        return {
-          label: v.name,
-          details: [{
-            label: `
-              <b>Type</b> ${v.type}<br>
-              <b>Repeating</b> ${v.repeating}<br>
-              <b>Comment</b> ${v.comment}<br>
-            `
-          }],
-          // children: [],
-          value: {
-            which: 'vars',
-            key: v.name.toLowerCase(),
-            value: v
-          }
-        }
-      })
-
-      const popupPages = []
-      // loop through the sorted pages once
-      const stepChildren = this.sortedPages.reduce((sc, page) => {
-        // capture popup pages into its own list
-        if (page.type === 'Popup') {
-          popupPages.push({
-            label: page.name,
-            details: [{
-              label: `
-                <b>Notes:</b> ${formatPageTextCell(page.notes)}<br>
-                <b>Text:</b> ${formatPageTextCell(page.text)}<br>
-              `
-            }],
-            // children: [],
-            value: {
-              which: 'pages',
-              key: page.name,
-              value: page
-            }
-          })
-          return sc
-        }
-        // else split the current page into its own array
-        if (!sc[page.step]) {
-          sc[page.step] = []
-        }
-        const fieldsDetail = {
-          label: 'Fields',
-          details: (page.fields || []).map(f => {
-            return {
-              label: `
-                <b>Type:</b> ${f.type}<br>
-                <b>Label:</b> ${f.label}<br>
-                <b>Variable:</b> ${f.name}<br>
-                <b>Invalid Prompt:</b> ${f.invalidPrompt}
-              `
-            }
-          })
-        }
-        const buttonsDetail = {
-          label: 'Buttons',
-          details: (page.buttons || []).map(b => {
-            return {
-              label: `
-                <b>Label:</b> ${b.label}<br>
-                <b>Next:</b> ${b.next}
-              `
-            }
-          })
-        }
-        const details = [{
-          label: `
-            <b>Notes:</b> ${formatPageTextCell(page.notes)}<br>
-            <b>Text:</b> ${formatPageTextCell(page.text)}<br>
-            <b>Learn More Prompt:</b> ${page.learn}<br>
-            <b>Help Text:</b> ${formatPageTextCell(page.help)}<br>
-            <b>Before Logic:</b> ${formatPageTextCell(page.codeBefore)}<br>
-            <b>After Logic:</b> ${formatPageTextCell(page.codeAfter)}<br>
-            <b>Logic Citation:</b> ${formatPageTextCell(page.codeCitation)}
-          `
-        }]
-        if (fieldsDetail.details.length) { details.push(fieldsDetail) }
-        if (buttonsDetail.details.length) { details.push(buttonsDetail) }
-        sc[page.step].push({
-          label: `Page ${page.name}`,
-          details,
-          // children: [],
-          value: {
-            which: 'pages',
-            key: page.name,
-            value: page
-          }
-        })
-        return sc
-      }, [])
-
-      accordion[1].children = guide.steps.map(s => {
-        const num = parseInt(s.number, 10)
-        return {
-          label: `Step ${num} - ${s.text}`,
-          details: [],
-          children: stepChildren[num],
-          value: {
-            which: 'steps',
-            key: num,
-            value: s
-          }
-        }
-      })
-
-      accordion[2].children = popupPages
-    }
-    // then media files
-    const media = this.guideFiles && this.guideFiles.media
-    if (media) {
-      accordion[3].children = media.map(file => {
-        return {
-          label: `${file.name}`,
-          details: [{
-            label: `
-            <b>Filename</b> ${file.filename}<br>
-            <b>Extension</b> ${file.extension}<br>
-            <b>Modified</b> ${file.modified}<br>
-            <b>Size</b> ${formatBytes(file.size)}<br>
-            `
-          }],
-          value: {
-            which: 'media',
-            key: file.filename,
-            value: file
-          }
-        }
-      })
-    }
-    // then template files
-    const templates = this.guideFiles && this.guideFiles.templates
-    if (templates) {
-      accordion[4].children = templates.map(template => {
-        return {
-          label: `${template.name}`,
-          details: [{
-            label: `
-            <b>Filename</b> ${template.filename}<br>
-            <b>Extension</b> ${template.extension}<br>
-            <b>Modified</b> ${template.modified}<br>
-            <b>Size</b> ${formatBytes(template.size)}<br>
-            `
-          }],
-          value: {
-            which: 'media',
-            key: template.filename,
-            value: template
-          }
-        }
-      })
-    }
-
-    return accordion
-  },
-
-  accordionValuesToGuidePartial (children, partial) {
-    children.forEach(checkedValueChildrenObj => {
-      // the 'value' from each item in accordionFromSourceGuide that was passed into the accordion
-      const checkedValue = checkedValueChildrenObj.value
-      // if we provided that value (the checkbox at the roots don't have values themselves in our case)
-      if (checkedValue) {
-        // destructure the 'value'
-        const { which, key, value } = checkedValue
-        // use the data from it to build up the partial guide
-        partial[which][key] = value // this 'value' is the actual (===) variable/step/page from the source guide
-      }
-      // if there are children here, recursively destructure their values and merge them into the partial too
-      if (checkedValueChildrenObj.children && checkedValueChildrenObj.children.length) {
-        this.accordionValuesToGuidePartial(checkedValueChildrenObj.children, partial)
-      }
-    })
-    // this partial will be merged into the target partial 1:1, depending on conflict resolutions
-    return partial
-  },
-
-  // convert the generic [{value, children[]}] output from the checkbox accordion back into a (partial) guide
-  guidePartialFromCheckedValues (checkedAccordionData) {
-    return this.accordionValuesToGuidePartial(checkedAccordionData, {
-      steps: {},
-      vars: {},
-      pages: {},
-      media: {},
-      templates: {}
-    })
-  }
-})
+import MergeToolGuideVM from './mege-tool-guide-vm'
 
 const canModelGuideListToPOJOArraySetter = v => {
   // targets and sources lists are set from app.stache, It's currently an old can-model type list
@@ -407,7 +40,7 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
   // search string that filters targets list
   targetsFilter: 'string',
   // on the left, clone of the target guide where the imported items are merging into
-  target: { default: () => new MergeToolGuide() },
+  target: { default: () => new MergeToolGuideVM() },
   // on the right, a list of source guides to show if one isn't selected yet
   sources: {
     default: () => [],
@@ -416,7 +49,7 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
   // search string that filters sources list
   sourcesFilter: 'string',
   // on the right, the selected guide from which we are checking boxes to bring its stuff into targetGuide
-  source: { default: () => new MergeToolGuide() },
+  source: { default: () => new MergeToolGuideVM() },
   // root of the recursive viewmodel pulled up from the accordion-checkbox component in this component's stache template
   sourceAccordionVM: 'any',
   // convert the generic [{value, children[]}] output from the checkbox accordion back into a (partial) guide
@@ -446,16 +79,32 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
   mergeStack: { default: () => [] },
   get mergedTarget () {
     const target = this.target && this.target.guide && this.target.guide.serialize()
+    let guideFiles = new DefineMap((this.target && this.target.guideFiles) || { media: [], templates: [] })
+    guideFiles = guideFiles.serialize()
+    
     const partials = this.mergeStack
     if (partials.length && target) {
       partials.forEach(p => {
         Object.assign(target.steps, p.steps)
         Object.assign(target.vars, p.vars)
         Object.assign(target.pages, p.pages)
+
+        ;['media', 'templates'].forEach(mt => {
+          const files = p[mt]
+          files && Object.keys(files).forEach(key => {
+            const file = (new DefineMap(files[key])).serialize()
+            const override = _findIndex(guideFiles[mt], f => (f.rename || f.filename) === (file.rename || file.filename))
+            if (override > -1) {
+              guideFiles[mt][override] = file
+            } else {
+              guideFiles[mt].push(file)
+            }
+          })
+        })
       })
     }
     target.title = this.mergeTitle || target.title
-    return new MergeToolGuide({ guide: new Guide(target) })
+    return new MergeToolGuideVM({ guide: new Guide(target), guideFiles })
   },
   mergeTitle: {
     type: 'string',
@@ -466,14 +115,7 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
     }
   },
 
-  // merge the selected values into the the target guide without any checks
-  simpleMerge () {
-    const sourceGuidePartial = this.sourceGuidePartial
-    this.mergeStack.push(sourceGuidePartial)
-    // this.interviewsPromise = undefined
-    this.cancelMerge()
-  },
-
+  // merge the selected values into the the target guide with minimal checks (overrides target)
   mergeSelected () {
     const sourceGuidePartial = this.sourceGuidePartial
     const { steps, pages } = sourceGuidePartial
@@ -528,6 +170,27 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
     const sourceGuidePartial = this.sourceGuidePartial
     const { vars, steps, pages } = sourceGuidePartial
     const mergedTarget = this.mergedTarget
+    const guideFiles = mergedTarget.guideFiles || { media: [], templates: [] }
+
+    // provide a safe 'rename' property on the file obj if the incoming filename collides
+    ;['media', 'templates'].forEach(mt => {
+      const importFiles = sourceGuidePartial[mt]
+      const importFilesList = importFiles && Object.keys(importFiles).map(key => importFiles[key])
+      const combinedList = guideFiles[mt].concat(importFilesList)
+      importFilesList.forEach(file => {
+        const fileNameOnly = file.filename.replace(/\d/g, '')
+        let fileNum = parseInt(file.filename.replace(/[^\d]/g, ''), 10) || 0
+        while (
+          _findIndex(
+            combinedList,
+            f => f !== file && (f.rename || f.filename) === (file.rename || file.filename)
+          ) > -1
+        ) {
+          file.rename = fileNameOnly + (fileNum++)
+        }
+      })
+    })
+    mergedTarget.guideFiles = guideFiles
 
     // Non-Overwriting Merge
     // update the sourceGuideParial so merging resolves as intended:
@@ -598,46 +261,47 @@ export const MergeToolVM = DefineMap.extend('MergeToolVM', {
       json: guideJsonStr
     }
 
-    let newGuideUrl
-    window.ws(saveNewGuideParams, function (data) {
+    let fileList = []
+    const templateIds = []
+    const gf = mergedTarget.guideFiles
+    if (gf) {
+      fileList = gf.media.concat(gf.templates).map(f => ({
+        filename: f.filename,
+        rename: f.rename || '',
+        extension: f.extension,
+        sourceGid: f.gid,
+        ownerfolder: f.ownerfolder
+      }))
+      gf.templates.forEach(t => templateIds.push((t.rename || t.filename).replace(/.*(\d+)$/, '$1')))
+    }
+
+    // return console.log(templateIds, fileList)
+
+    return window.ws(saveNewGuideParams, function (data) {
       if (data.error !== undefined) {
         // TODO: setProgress(data.error)
         errors.push(data.error)
         console.error(data.error)
       } else {
         const newgid = data.gid // new guide id
-        newGuideUrl = data.url // location of new guide for file copying
+        // const newGuideUrl = data.url // location of new guide for file copying
         guide.id = newgid
         if (typeof vm.reloadInterviews === 'function') {
           vm.reloadInterviews()
         }
 
-        // TODO: successful merge process should be promise chain?
-        // Also, this currently only copies media, not templates
-        const list = []
-        vm.mergeStack.forEach(stackItem => {
-          const media = stackItem.media
-          const keys = media && Object.keys(media)
-          if (keys.length) {
-            keys.forEach(key => {
-              const fileToCopy = media[key]
-              const filePath = fileToCopy.path
-              const fileName = fileToCopy.name
-              if (filePath) {
-                list.push({ fileName, filePath })
-              }
-            })
-          }
-        })
-
-        // copy the list of files
-        window.ws({
-          cmd: 'copyfiles',
-          fileList: list,
-          targetPath: newGuideUrl
-        }, function (data) {
-          console.log(data)
-        })
+        if (fileList.length) {
+          // copy the list of files
+          // TODO: promise chain return, block UI while saving
+          window.ws({
+            cmd: 'copyfiles',
+            gid: newgid,
+            fileList,
+            templateIds
+          }, function (data) {
+            console.log(newgid, data)
+          })
+        }
 
         // return to merge view
         vm.open = false
@@ -678,6 +342,7 @@ export default Component.extend({
       const sizeInKB = Math.ceil(sizeInBytes / 1024)
       return sizeInKB ? `${sizeInKB}K` : ''
     },
-    formatPageTextCell
+    formatPageTextCell,
+    formatBytes
   }
 })
